+++ packages/client/src/devices/__tests__/mocks.ts
@@ -5,7 +5,8 @@ import {
   OwnCapability,
 } from '../../gen/coordinator';
 import { Call } from '../../Call';
-import { Subject } from 'rxjs';
+import { of, Subject } from 'rxjs';
+import { BrowserPermission } from '../BrowserPermission';
 
 export const mockVideoDevices = [
   {
@@ -217,3 +218,7 @@ export const mockDeviceIds$ = () => {
 export const emitDeviceIds = (values: MediaDeviceInfo[]) => {
   deviceIds.next(values);
 };
+
+export const mockBrowserPermission = {
+  asObservable: () => of(true),
+} as BrowserPermission;