+++ packages/react-native-sdk/jest.config.ts
@@ -4,7 +4,8 @@ const config: Config = {
   preset: 'react-native',
   setupFilesAfterEnv: [
     '<rootDir>/jest-setup.ts',
-    '<rootDir>/node_modules/react-native-gesture-handler/jestSetup.js',
+    // this package is hoisted to the root node_modules
+    '<rootDir>/../../node_modules/react-native-gesture-handler/jestSetup.js',
   ],
   testPathIgnorePatterns: [
     '<rootDir>/__tests__/mocks/',