+++ packages/react-sdk/src/components/CallControls/ToggleVideoButton.tsx
@@ -31,6 +31,7 @@ export const ToggleVideoPreviewButton = (
     caption,
     Menu = DeviceSelectorVideo,
     menuPlacement = 'top',
+    onMenuToggle,
     ...restCompositeButtonProps
   } = props;
   const { t } = useI18n();
@@ -64,8 +65,11 @@ export const ToggleVideoPreviewButton = (
         disabled={!hasBrowserPermission}
         Menu={Menu}
         menuPlacement={menuPlacement}
-        onMenuToggle={(shown) => setTooltipDisabled(shown)}
         {...restCompositeButtonProps}
+        onMenuToggle={(shown) => {
+          setTooltipDisabled(shown);
+          onMenuToggle?.(shown);
+        }}
       >
         <Icon icon={!optimisticIsMute ? 'camera' : 'camera-off'} />
         {!hasBrowserPermission && (
@@ -95,6 +99,7 @@ export const ToggleVideoPublishingButton = (
     caption,
     Menu = <DeviceSelectorVideo visualType="list" />,
     menuPlacement = 'top',
+    onMenuToggle,
     ...restCompositeButtonProps
   } = props;
 
@@ -153,8 +158,11 @@ export const ToggleVideoPublishingButton = (
             Menu={Menu}
             menuPlacement={menuPlacement}
             menuOffset={16}
-            onMenuToggle={(shown) => setTooltipDisabled(shown)}
             {...restCompositeButtonProps}
+            onMenuToggle={(shown) => {
+              setTooltipDisabled(shown);
+              onMenuToggle?.(shown);
+            }}
           >
             <Icon icon={optimisticIsMute ? 'camera-off' : 'camera'} />
             {(!hasBrowserPermission ||