+++ sample-apps/react/egress-composite/src/CompositeApp.tsx
@@ -23,6 +23,11 @@ import { useParticipantStyles } from './hooks/options/useParticipantStyles';
 export const CompositeApp = () => {
   const { client, call } = useInitializeClientAndCall();
 
+  // @ts-expect-error makes it easy to debug in the browser console
+  window.call = call;
+  // @ts-expect-error makes it easy to debug in the browser console
+  window.client = client;
+
   return (
     <StreamVideo client={client}>
       <StreamCall call={call}>